[crypto] Rename lib/crypto/impl/rsa_3072 to lib/crypto/impl/rsa.
This will ultimately result in less clutter in the impl/ directory by
keeping all RSA-related code together.
