+++ sw/device/lib/crypto/impl/rsa/rsa_3072_verify.h
@@ -2,8 +2,8 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef OPENTITAN_SW_DEVICE_LIB_CRYPTO_IMPL_RSA_3072_RSA_3072_VERIFY_H_
-#define OPENTITAN_SW_DEVICE_LIB_CRYPTO_IMPL_RSA_3072_RSA_3072_VERIFY_H_
+#ifndef OPENTITAN_SW_DEVICE_LIB_CRYPTO_IMPL_RSA_RSA_3072_VERIFY_H_
+#define OPENTITAN_SW_DEVICE_LIB_CRYPTO_IMPL_RSA_RSA_3072_VERIFY_H_
 
 #include <stddef.h>
 #include <stdint.h>
@@ -158,4 +158,4 @@ status_t rsa_3072_verify(const rsa_3072_int_t *signature,
 }  // extern "C"
 #endif  // __cplusplus
 
-#endif  // OPENTITAN_SW_DEVICE_LIB_CRYPTO_IMPL_RSA_3072_RSA_3072_VERIFY_H_
+#endif  // OPENTITAN_SW_DEVICE_LIB_CRYPTO_IMPL_RSA_RSA_3072_VERIFY_H_
