+++ sw/device/tests/crypto/BUILD
@@ -154,7 +154,7 @@ opentitan_functest(
         ":rsa_3072_verify_testvectors_wycheproof_header",
         "//sw/device/lib/crypto/drivers:hmac",
         "//sw/device/lib/crypto/drivers:otbn",
-        "//sw/device/lib/crypto/impl/rsa_3072:rsa_3072_verify",
+        "//sw/device/lib/crypto/impl/rsa:rsa_3072_verify",
         "//sw/device/lib/runtime:log",
         "//sw/device/lib/testing/test_framework:ottf_main",
     ],
@@ -177,7 +177,7 @@ opentitan_functest(
         ":rsa_3072_verify_testvectors_hardcoded_header",
         "//sw/device/lib/crypto/drivers:hmac",
         "//sw/device/lib/crypto/drivers:otbn",
-        "//sw/device/lib/crypto/impl/rsa_3072:rsa_3072_verify",
+        "//sw/device/lib/crypto/impl/rsa:rsa_3072_verify",
         "//sw/device/lib/runtime:log",
         "//sw/device/lib/testing/test_framework:ottf_main",
     ],
